Implemented Game::doRound and helpers.
The doRound function returns the successfulness of the defend and
otherwise just does a round against a single defender.  Uses the helper
function getAttackingCard which gets the next attackingCard even through
passes.  This function in turn uses the nextAttacker helper function to
safely move the attacker while wrapping around and skipping the current
defender.