Changed the GameListener interface and implemented some messages.
AI player is no longer a GameListener as it didn't do anythign with the
messages anyways!  Players still get an agent pointer at the start of
the game so they can do queries.

Many of the GameListener messages are implemented.
Messages that still need to be implemented:
GameListener::attackerChanged()
GameListener::givenCards()