Fixed segfault in game. Updated AI.
Again with nextDefender shenanigans.  Revert previous "bugfix".  Added
sanity checks(asserts) to trip before the segfault at least and give a
better context for a debugger.

The AI is now quite a bit smarter.  It doesn't just play randomly
anymore.  It attacks and defends with it's lowest card and doesn't pile
on trump.