#31 Removed SynchronizationAnimation and replaced with CrunchAnimation.
This fixes a hang when the game_thread is waiting on the synchronization aniamtiont to complete in gameOver or givenCards, etc.  CrunchAnimation has the same effect of waiting for the proper time to create the animation but now does it explicitly instead of with thread synchronization.

Also created the Animation.cpp and put the CrunchAnimation definitions there (starting #32)